+++ instrumentation/opentelemetry-instrumentation-flask/tests/test_automatic.py
@@ -59,3 +59,23 @@ class TestAutomatic(InstrumentationTest, TestBase, WsgiTestBase):
         self.assertEqual([b"Hello: 123"], list(resp.response))
         span_list = self.memory_exporter.get_finished_spans()
         self.assertEqual(len(span_list), 1)
+
+    def test_exluded_urls_explicit(self):
+        FlaskInstrumentor().uninstrument()
+        FlaskInstrumentor().instrument(excluded_urls="/hello/456")
+
+        self.app = flask.Flask(__name__)
+        self.app.route("/hello/<int:helloid>")(self._hello_endpoint)
+        client = Client(self.app, BaseResponse)
+
+        resp = client.get("/hello/123")
+        self.assertEqual(200, resp.status_code)
+        self.assertEqual([b"Hello: 123"], list(resp.response))
+        span_list = self.memory_exporter.get_finished_spans()
+        self.assertEqual(len(span_list), 1)
+
+        resp = client.get("/hello/456")
+        self.assertEqual(200, resp.status_code)
+        self.assertEqual([b"Hello: 456"], list(resp.response))
+        span_list = self.memory_exporter.get_finished_spans()
+        self.assertEqual(len(span_list), 1)